requirements.txt
blinker==1.9.0
click==8.2.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
Werkzeug==3.1.3
Flask
psycopg2-binary
